requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-Limiter==3.5.0
Werkzeug==3.0.1
g4f==0.2.0
beautifulsoup4==4.12.2
requests==2.31.0
APScheduler==3.10.4
python-dotenv==1.0.1
SQLAlchemy==2.0.25
Flask-Migrate==4.0.5
PyJWT==2.8.0
psycopg2-binary==2.9.9
redis==5.0.1 